Bug #1675724 unisoc: wlan: modify variable type
[root cause  ] modify variable type
[changes     ] drivers/unisoc_platform/wlan/
[side effects] none
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:none
[confirm dependent commit]:none
[board]:sharkl5pro
[test case]:none
[reviewers   ] baolei.yuan

[change_type ] bringup --> sprd_chip_support
[tag_product ] common

+++ drivers/unisoc_platform/wlan/common/cfg80211.h
@@ -239,7 +239,7 @@ struct sprd_sched_scan {
 	u32 interval;
 	u32 flags;
 	s32 rssi_thold;
-	u8 channel[SPRD_TOTAL_CHAN_NR];
+	u16 channel[SPRD_TOTAL_CHAN_NR];
 
 	u32 n_ssids;
 	u8 *ssid[SPRD_TOTAL_CHAN_NR];
